requirements.txt
amqp==2.6.1
anyjson==0.3.3
asgiref==3.2.10
billiard==3.6.4.0
celery==4.4.2
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.4
Django==3.1
django-celery-beat==2.2.0
django-timezone-field==4.1.2
djangorestframework==3.12.0
djangorestframework-simplejwt==4.4.0
dnspython==1.16.0
eventlet==0.29.1
greenlet==1.1.0
importlib-metadata==4.0.1
kombu==4.6.11
prompt-toolkit==3.0.18
pyclean==2.0.0
PyJWT==2.1.0
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
vine==1.3.0
wcwidth==0.2.5
zipp==3.4.1
